+++ resources/org/uecide/changelog.md
@@ -4,6 +4,7 @@ Changelog
 0.10.6
 ------
 
+* Added Material LAF and cleaned up icons and buttons
 * Complete overhaul of the icon system;
 * Added changing icon for toggle button
 * Add sketch name to serial terminal title#
@@ -680,6 +681,7 @@ Changelog
 0.8.5
 -----
 
+* Added Material LAF and cleaned up icons and buttons
 * Complete overhaul of the icon system;
 * Added changing icon for toggle button
 * Add sketch name to serial terminal title#